requirements.txt
asgiref==3.3.1
Django==3.1.6
djangorestframework==3.12.2
pytz==2019.2
sqlparse==0.4.1
